Complete the implementation of useFetchAndCombine
Create state variables for data and loading. Filter results
of Promise.all returns in such a way that all the media is
first sorted and only 5-7 for each are spliced out and
pushed into a new array, which get update to the data variable.

Check the result of this implementation in SearchPage component.