Add endpoint in new hook and create new card components
The useFetchAndCombine() hooks returns data in which
each item doesn't have a media_type or endpoint property
that is used across this app in Carousel, MediaCard
components to navigate to that particular item details page.

Add endpoint/mediaType to each item that is returned in data
from the useFetchAndCombine() hook.

Create a new card component that will be rendered at the SearchPage
for genre-wise results only.